requirements.txt
# ---------- Core API ----------
fastapi==0.128.4
uvicorn==0.40.0
starlette==0.52.1

# ---------- ORM / DB ----------
SQLAlchemy==2.0.46
psycopg==3.3.2

# ---------- Elasticsearch (MATCH ES SERVER v8) ----------
elasticsearch>=8,<9
elastic-transport>=8,<9

# ---------- AWS ----------
boto3==1.42.44
botocore==1.42.44
s3transfer==0.16.0

# ---------- Google Cloud ----------
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-api-core==2.29.0
google-auth==2.48.0
google-resumable-media==2.8.0
google-crc32c==1.8.0
googleapis-common-protos==1.72.0
proto-plus==1.27.1
protobuf==6.33.5

# ---------- Firebase ----------
firebase-admin==7.1.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0

# ---------- Config / HTTP ----------
python-dotenv==1.2.1
requests==2.32.5
httpx==0.28.1

# ---------- Security ----------
cryptography==46.0.4
PyJWT==2.11.0

# ---------- Utils ----------
python-dateutil==2.9.0.post0
typing_extensions==4.15.0
packaging==26.0



